This change allows Play! to pass parameters that lack values to the Controller method. e.g. /servlet?foo&bar&baz would not previously make 'foo', 'bar', or 'baz' available to the receiving method.
This change solves that problem, binding value-less parameter names to null. That is indistinguishable from "no param present" if the method receives it as direct parameter, but these value-less parameters can be accessed by using params.all() or params.allSimple() methods to get a Map.

A selenium test is included